Sign-up page / Log-in page
This project, you will build a sign-up or log-in page with the necessary input fields for a username and a password.
Because we can create a user profile on almost every website, forms are absolutely essential for allowing people to set their usernames and passwords.Your forms will collect inputs from users and a separate back-end program will know how to store and process that data.Creating a clean and clear sign-up page can be surprisingly difficult. The more you learn about HTML and CSS, the more content you want to create to showcase your skills. But the thing is: a sign-up page needs to be as clean and easy-to-use as possible.Thus, the biggest challenge with this project is to keep it simple, clear, and light.